requirements.txt
Flask==2.0.1
gunicorn==20.1.0
Flask-CORS==3.0.10
Flask-Limiter==2.6.0
python-dotenv==0.19.0
